==> блокноты по откртытым данным/1 Linear Model.ipynb <==
import pandas as pd
import numpy as np
df = pd.read_csv("data.csv")
df.head()

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, classification_report, roc_auc_score
from sklearn.preprocessing import StandardScaler

#Создание для второго теста датасета без столба CourseCategory
df2 = df.drop(columns=['CourseCategory'])

#One Hot Encoding
zeros_array = np.zeros_like(df['UserID'])
df['is_health'] = zeros_array
df['is_health'] = df['is_health'].mask(df['CourseCategory'] == 'Health' , 1)
df['is_arts'] = zeros_array
df['is_arts'] = df['is_arts'].mask(df['CourseCategory']== 'Arts' , 1)
df['is_science'] = zeros_array
df['is_science'] = df['is_science'].mask(df['CourseCategory']== 'Science' , 1)
df['is_programming'] = zeros_array
df['is_programming'] = df['is_programming'].mask(df['CourseCategory']== 'Programming' , 1)
df.head()

df = df.drop(columns=['CourseCategory'])
df.columns

df.head(1)

df.to_csv('data_One_hot.csv', index=False)

X = df.drop('CourseCompletion', axis=1)
y = df['CourseCompletion']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=50)
scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

model = LogisticRegression(solver='liblinear', random_state=42) 

model.fit(X_train, y_train)

y_pred = model.predict(X_test)

accuracy = accuracy_score(y_test, y_pred)
print(f"Accuracy: {accuracy}")
print(classification_report(y_test, y_pred))
print(f"AUC-ROC: {roc_auc_score(y_test, model.predict_proba(X_test)[:, 1])}")

df2.head()

X = df2.drop('CourseCompletion', axis=1)
y = df2['CourseCompletion']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=50)
scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)
model = LogisticRegression(solver='liblinear', random_state=42) 
model.fit(X_train, y_train)
y_pred = model.predict(X_test)
accuracy = accuracy_score(y_test, y_pred)
print(f"Accuracy: {accuracy}")
print(classification_report(y_test, y_pred))
print(f"AUC-ROC: {roc_auc_score(y_test, model.predict_proba(X_test)[:, 1])}")

#Интересное наблюдение, по сути нет рзницы между наличием и отсутствием признака вид курсов.

df2.to_csv('data_notypes.csv', index=False)

df3 = df.drop(columns=['UserID'])

df3.head()

#Проба вообще убрать ID как в KNN.
#Не улучшает ситуации. Видимо, модель ей придает маленький вес.

X = df3.drop('CourseCompletion', axis=1)
y = df3['CourseCompletion']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=50)
scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)
model = LogisticRegression(solver='liblinear', random_state=42) 
model.fit(X_train, y_train)
y_pred = model.predict(X_test)
accuracy = accuracy_score(y_test, y_pred)
print(f"Accuracy: {accuracy}")
print(classification_report(y_test, y_pred))
print(f"AUC-ROC: {roc_auc_score(y_test, model.predict_proba(X_test)[:, 1])}")
