==> блокноты по откртытым данным/2 knn Model.ipynb <==
import pandas as pd
import numpy as np
df = pd.read_csv("data_notypes.xls")
df = df.drop(columns=['UserID'])
df.head()

from sklearn.model_selection import train_test_split
from sklearn.neighbors import KNeighborsClassifier, KNeighborsRegressor
from sklearn.metrics import accuracy_score, mean_squared_error
from sklearn.preprocessing import StandardScaler

X = df.drop('CourseCompletion', axis=1)
X = X.drop('DeviceType', axis=1)
y = df['CourseCompletion']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=50)
scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

model_class = KNeighborsClassifier(n_neighbors=3)
model_class.fit(X_train, y_train)
y_pred = model_class.predict(X_test)
accuracy = accuracy_score(y_test, y_pred)
print(f"Accuracy (Classifier): {accuracy}")

from sklearn.metrics import roc_auc_score
roc_auc = roc_auc_score(y_test, y_pred)

print(f"ROC AUC: {roc_auc}")

model_reg = KNeighborsRegressor(n_neighbors=3)
model_reg.fit(X_train, y_train)
y_pred = model_reg.predict(X_test)
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error (Regressor): {mse}")

#Первый

df = pd.read_csv("data_One_hot.csv")
df = df.drop(columns=['UserID'])
df.head()

X = df.drop('CourseCompletion', axis=1)
y = df['CourseCompletion']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=50)
scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

model_class = KNeighborsClassifier(n_neighbors=3)
model_class.fit(X_train, y_train)
y_pred = model_class.predict(X_test)
accuracy = accuracy_score(y_test, y_pred)
print(f"Accuracy (Classifier): {accuracy}")

model_reg = KNeighborsRegressor(n_neighbors=3)
model_reg.fit(X_train, y_train)
y_pred = model_reg.predict(X_test)
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error (Regressor): {mse}")

#Второй

import pandas as pd
import numpy as np
df = pd.read_csv("data.csv")
df = df.drop(columns=['UserID'])
df.head()

df['Type'] = np.nan
df['Type'] = df['Type'].mask(df['CourseCategory'] == 'Health' , 1)
df['Type'] = df['Type'].mask(df['CourseCategory'] == 'Arts', 2)
df['Type'] = df['Type'].mask(df['CourseCategory'] == 'Science', 3)
df['Type'] = df['Type'].mask(df['CourseCategory'] == 'Programming', 4)
df['Type'] = df['Type'].mask(df['CourseCategory'] == 'Business', 5)
df.head(5)

df = df.drop(columns=['CourseCategory'])

df.head(5)

X = df.drop('CourseCompletion', axis=1)
y = df['CourseCompletion']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=50)
scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

model_class = KNeighborsClassifier(n_neighbors=3)
model_class.fit(X_train, y_train)
y_pred = model_class.predict(X_test)
accuracy = accuracy_score(y_test, y_pred)
print(f"Accuracy (Classifier): {accuracy}")

model_reg = KNeighborsRegressor(n_neighbors=3)
model_reg.fit(X_train, y_train)
y_pred = model_reg.predict(X_test)
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error (Regressor): {mse}")

#Лучше всего вообще убрать столбец тип курса, как оказалось.

df = pd.read_csv("data_notypes.csv")
X = df.drop('CourseCompletion', axis=1)
y = df['CourseCompletion']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=50)
scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)


model_class = KNeighborsClassifier(n_neighbors=3)
model_class.fit(X_train, y_train)
y_pred = model_class.predict(X_test)
accuracy = accuracy_score(y_test, y_pred)
print(f"Accuracy (Classifier): {accuracy}")

model_reg = KNeighborsRegressor(n_neighbors=3)
model_reg.fit(X_train, y_train)
y_pred = model_reg.predict(X_test)
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error (Regressor): {mse}")

#Последний пример, показывает, что нужно убрать столбец ID!
