==> блокноты по откртытым данным/3 Desicion tree .ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from catboost import CatBoostClassifier, Pool
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix

pip install catboost

df = pd.read_csv("data.csv")
df.head()

df.columns

X = df.drop('CourseCompletion', axis=1)
y = df['CourseCompletion']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=50)

#1 глубокое дерево
categorical_features = ['CourseCategory']
train_pool = Pool(data=X_train, label=y_train, cat_features=categorical_features)
test_pool = Pool(data=X_test, label=y_test, cat_features=categorical_features)
model = CatBoostClassifier(iterations=1, random_state=42, verbose=0, depth=16)
model.fit(train_pool)
y_pred = model.predict(test_pool)

accuracy = accuracy_score(y_test, y_pred)
print(f"Accuracy: {accuracy}")
print("Classification Report:")
print(classification_report(y_test, y_pred))
print("Confusion Matrix:")
print(confusion_matrix(y_test, y_pred))

#10 деревьев глубины 3
categorical_features = ['CourseCategory']
train_pool = Pool(data=X_train, label=y_train, cat_features=categorical_features)
test_pool = Pool(data=X_test, label=y_test, cat_features=categorical_features)
model = CatBoostClassifier(iterations=10, random_state=42, verbose=0, depth=3)
model.fit(train_pool)
y_pred = model.predict(test_pool)

accuracy = accuracy_score(y_test, y_pred)
print(f"Accuracy: {accuracy}")
print("Classification Report:")
print(classification_report(y_test, y_pred))
print("Confusion Matrix:")
print(confusion_matrix(y_test, y_pred))

#100 деревьев глубины 4
categorical_features = ['CourseCategory']
train_pool = Pool(data=X_train, label=y_train, cat_features=categorical_features)
test_pool = Pool(data=X_test, label=y_test, cat_features=categorical_features)
model = CatBoostClassifier(iterations=100, random_state=42, verbose=0, depth=4)
model.fit(train_pool)
y_pred = model.predict(test_pool)

accuracy = accuracy_score(y_test, y_pred)
print(f"Accuracy: {accuracy}")
from sklearn.metrics import roc_auc_score
roc_auc = roc_auc_score(y_test, y_pred)
print(f"ROC AUC: {roc_auc}")

model.save_model("model.cbm")

#Проверим без ID. Вдруг человек с определенным ID - умный. И модель видя его и в теренировочных и в тестовых данных понимает
#Что он точно завершит курс. И наоборот в случае, если он не умный
df = df.drop(columns=['UserID'])

X = df.drop('CourseCompletion', axis=1)
y = df['CourseCompletion']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=50)

#30 деревьев глубины 3
categorical_features = ['CourseCategory']
train_pool = Pool(data=X_train, label=y_train, cat_features=categorical_features)
test_pool = Pool(data=X_test, label=y_test, cat_features=categorical_features)
model = CatBoostClassifier(iterations=30, random_state=42, verbose=0, depth=3)
model.fit(train_pool)
y_pred = model.predict(test_pool)

accuracy = accuracy_score(y_test, y_pred)
print(f"Accuracy: {accuracy}")
print("Classification Report:")
print(classification_report(y_test, y_pred))
print("Confusion Matrix:")
print(confusion_matrix(y_test, y_pred))

# Результат чуть хуже, но не сильно. Возможно это связано с информативностью столбика ID, а возможно с предположением описаным выше.
# Этот столбец мог помогть, так как через него данные из train частично просачиваются в test
